Converting opened field from int to int8

+++ jobs/etl/photosphere/ab_photosphere.py
@@ -1,13 +1,14 @@
 # coding=utf-8
 
+import os
+import sys
+
 import boto3
 import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pyarrow.parquet as pq
 
-import os
-import sys
 here = os.path.dirname(os.path.realpath(__file__))
 sys.path.append(os.path.join(here, '../../../'))
 from jobs.wrappers.amplitude.amplitude_athena_wrapper import AthenaAmplitudeETL
@@ -133,7 +134,7 @@ usability_schema = [('eventdate', np.str),
                     ('imovel_id', np.int64),
                     ('photosphere_id', np.str),
                     ('viewed', np.int8),
-                    ('opened', np.int)]
+                    ('opened', np.int8)]
 create_parquet(usability_key, metrics['usability'], usability_schema)
 
 funnel_key = 'clean/amplitude/ab_tests/photosphere/funnel_conversion/funnel_conversion.parq'
